Organized project structure into folders
- Dependencies, Data, Categories, Models, Printers, Interfaces
