+++ TheatricalPlayersRefactoringKata.Tests/XmlStatementPrinterTests.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Text;
 using System.Xml;
+using TheatricalPlayersRefactoringKata.Models;
 
 namespace TheatricalPlayersRefactoringKata
 {
