+++ TheatricalPlayersRefactoringKata/Categories/TragedyCategory.cs
