+++ TheatricalPlayersRefactoringKata/Interfaces/IPlayCategory.cs
